(gh-21) Update Gemfile to pull in openvox gem instead of puppet
I haven't yet tracked down exactly what the puppet gem dependencies are
within this project, but this just switches to using openvox instead so
that we aren't depending/testing on the older, dead puppetlabs branch.

By default, puppet_ref is 'main' and bundler will pull in openvox@main
as a gem.