Remove update checking code
By default, PuppetDB check the current version against the latest
version available and log an information message when a new version is
available.

We do not have a similar system in place, and might even consider this
to not be aligned with our values, so remove version checking code and
also the configuration option that allowed to disable it.