(gh-21) Add a step to run beaker-puppet configure_type_defaults_on()
This completes the .ssh/environment setup for beaker. Normally beaker-puppet
would do this as part of its installation steps, but it's not capable of
installing openvox packages, so we will install the packages ahead of time
before Beaker gets involved.

The configure_type_defaults_on() step is important because
Beaker's use of ssh as root relies on the .ssh/environment file
supplying PATH to the puppet binaries.