(gh-21) Drop the firewalls step from the acceptance presuite
The iptables command isn't present anymore in el8+, which uses nft. This
pre-suite step is an ancient one that was put in place for ec2 and
vagrant testing, neither of which applies for the kvm acceptance testing we
are doing in the github action, so I'm just skipping it rather than
trying to make it firewall agnostic.