(gh-21) Abstract direct calls to sysv service
There are a few spots in the suite where an attempt is made to 'service
foo action' something directly, usually because we're testing status
code or attempting to reload. EL9 and others no longer have service, so
I've added a little service() helper to lib/helper.rb to pick systemctl
or fallback to service or squawk.

Where it appeared that puppet-resource would suffice, I switched to
using that.

However a little deeper in the rabbit hole lurks beaker-puppet which
provides the with_puppet_running_on() method used throughout the suite,
which calls a bounce_service method also in
BeakerPuppet::Helpers::PuppetHelpers. And that method *also* attempts to
use the sysv service executable...

Right now I'm overriding this locally in helpers.rb to sidestep the
issue, but we may want to bite the bullet, fork beaker-puppet and either
depend on the forked repo directly or generate a different gem.

Have not yet bitten the bullet. This is the same workaround I have in:
openvoxproject/openvox-server@4758fec54d6a
openvoxproject/openvox@2f85330b84d236f
...