(gh-21) Remove unused beaker plugins from acceptance dependencies
beaker-abs should only be relevant inside perforce.
Someone could conceivably have setup their own vmpooler, and want to run
the acceptance suite using it I suppose, but we could always add it back
if we had to.
Likewise an explicit faraday dependency is probably no longer needed, as
that was added for an old puppet dependency in a8ac8bf6.