Update report update condition to use report ID
When multiple identical reports are submitted, there can be a race condition where a report is committed for that node during the time the second report is being processed and they are both trying to be set as the latest report, resulting in a duplicate report being stored. This causes any subsequent report submissions to fail for that node. This fixes the problem by using the latest-report-id calculated when the function is called, using its timestamp to decide if this report is in fact the latest report.

Fixes #16.